Replace en-dash with hyphen in diff output for better Git compatibility
This commit replaces the en-dash (–) with a hyphen (-) in the failStringsEqualWithDiff function. The reason for this change is to improve compatibility with the Git diff standard, which always uses a hyphen to indicate removed code.

This change is particularly useful when the message generated by the assert function is passed to GitHub or Swift Forums posts. Wrapping the message in a diff-code-block now requires no additional adjustments to the symbol generated by our function.